Generalize barcode capture widget
