+++ rdt-open-reader/src/covidIndonesia/assets/json.form-id/sample-collection-form-original.json
@@ -543,7 +543,7 @@
         "openmrs_entity_parent": "",
         "openmrs_entity": "",
         "openmrs_entity_id": "",
-        "type": "barcode",
+        "type": "covid_rdt_barcode_reader",
         "barcode_type": "qrcode",
         "hint": "",
         "scanButtonText": "",
